requirements.txt
django==4.1.11
djongo==1.3.6
pymongo==3.12.3
sqlparse==0.2.4
djangorestframework==3.12.2
djangorestframework-simplejwt==5.3.0
django-cors-headers==3.7.0
django-environ==0.11.2
pyjwt==2.8.0
pytz